requirements.txt
streamlit
pandas
Pillow
numpy